[chore] 카메라 줌 아웃에 따라 화면 이동속도 변경

+++ Assets/Scripts/Contents/Placement/Place/PlacementObjectInfo.cs
@@ -26,7 +26,7 @@ public class PlacementLevelInfo
     [field: SerializeField]
     public string Name { get; private set; }
     [field: SerializeField]
-    public int DefaultHp { get; private set; }
+    public float DefaultHp { get; private set; }
     [field: SerializeField]
     public GameObject Prefeb { get; private set; }
     [field: SerializeField]
